add merge speed
